refactor: search bar dep in array
