+++ src/hooks/useHomeFetch.ts
@@ -1,28 +1,18 @@
 import { useEffect, useState } from "react";
 import { api } from "../API";
-import { MoviePropTypes } from "../Global.props";
 import { getPersistedState } from "../helpers";
-
-/**
- * Represents the initial state for the home fetch hook.
- */
-const initialState = {
-  page: 0,
-  results: [] as MoviePropTypes[],
-  total_pages: 0,
-  total_results: 0,
-};
+import { initialState } from "./props";
 
 /**
  * Custom hook for fetching movies for the home page.
  * @returns An object containing the state, loading status, error status, search term, setSearchTerm function, and isLoadingMore status.
  */
 function useHomeFetch() {
-  const [error, setError] = useState(false);
-  const [isLoadingMore, setIsLoadingMore] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [searchTerm, setSearchTerm] = useState("");
-  const [state, setState] = useState(initialState);
+  const [ error, setError ] = useState(false);
+  const [ isLoadingMore, setIsLoadingMore ] = useState(false);
+  const [ loading, setLoading ] = useState(false);
+  const [ searchTerm, setSearchTerm ] = useState("");
+  const [ state, setState ] = useState(initialState);
 
   /**
    * Fetches movies from the API based on the specified page and search term.
@@ -40,8 +30,8 @@ function useHomeFetch() {
         ...movies,
         results:
           page > 1
-            ? [...prev.results, ...(movies.results || [])]
-            : [...(movies.results || [])],
+            ? [ ...prev.results, ...(movies.results || []) ]
+            : [ ...(movies.results || []) ],
       }));
     } catch (error) {
       setError(true);
@@ -63,17 +53,17 @@ function useHomeFetch() {
     console.log("Grabbing from API");
     setState(initialState);
     fetchMovies(1, searchTerm);
-  }, [searchTerm]);
+  }, [ searchTerm ]);
 
   useEffect(() => {
     if (!isLoadingMore) return;
     fetchMovies(state.page + 1, searchTerm);
     setIsLoadingMore(false);
-  }, [isLoadingMore, searchTerm, state.page]);
+  }, [ isLoadingMore, searchTerm, state.page ]);
 
   useEffect(() => {
     if (!searchTerm) sessionStorage.setItem("homeState", JSON.stringify(state));
-  }, [searchTerm, state]);
+  }, [ searchTerm, state ]);
 
   return { state, loading, error, searchTerm, setSearchTerm, setIsLoadingMore };
 }
